+++ android/src/test/kotlin/com/project/starter/quality/AndroidQualityPluginTest.kt
@@ -62,7 +62,7 @@ internal class AndroidQualityPluginTest : WithGradleProjectTest() {
                     
                     repositories {
                         google()
-                        jcenter()
+                        mavenCentral()
                     }
                     
                     android {