+++ jvm/src/main/kotlin/com/project/starter/modules/internal/Repositories.kt
@@ -3,23 +3,23 @@ package com.project.starter.modules.internal
 import org.gradle.api.Project
 
 internal fun Project.configureRepositories(): Unit = with(repositories) {
-    google { repository ->
-        repository.mavenContent { content ->
+    exclusiveContent { content ->
+        content.forRepository { google() }
+        content.filter {
             val googleLibraries = listOf(
                 "com\\.android.*",
                 "androidx.*",
                 "android.arch.*",
                 "com\\.google.*"
             )
-            googleLibraries.forEach(content::includeGroupByRegex)
+            googleLibraries.forEach(it::includeGroupByRegex)
         }
+    }
+    exclusiveContent { content ->
+        content.forRepository {
+            maven { it.setUrl("https://oss.sonatype.org/content/repositories/snapshots") }
+        }
+        content.filter { it.includeGroup("org.jacoco") }
     }
     mavenCentral()
-    @Suppress("DEPRECATION")
-    jcenter { repository ->
-        repository.mavenContent { content ->
-            content.includeGroup("org.jetbrains.trove4j")
-            content.includeGroup("org.jetbrains.kotlinx")
-        }
-    }
 }