+++ multiplatform/src/main/kotlin/com/project/starter/modules/internal/MultiplatformCoverage.kt
@@ -18,7 +18,7 @@ internal fun Project.configureMultiplatformCoverage() {
     }
 
     extensions.configure(JacocoPluginExtension::class.java) {
-        it.toolVersion = "0.8.6"
+        it.toolVersion = "0.8.7-SNAPSHOT"
     }
     tasks.register("jacocoTestReport", JacocoReport::class.java) {
         it.dependsOn(":$path:jvmTest")